Documents/streamlit_judge_sheet_condition_app/requirements.txt
numpy==1.26.0
opencv-contrib-python-headless==4.8.1.78
pillow==10.1.0
rembg==2.0.50
onnxruntime==1.16.0
PyYAML==6.0.1
pandas==2.1.1
scikit-learn==1.3.1
pytz==2023.3